requirements.txt
alembic
fastapi
pydantic
python-dotenv
Requests
SQLAlchemy
python-jose[cryptography]
psycopg-binary
uvicorn
httpx
